Rename import_handler module.
Rename the module and corresponding test module; the
contents of the module have changed significantly. At
this point, that module is essentially capturing the
optional dependencies used during testing.
