+++ traits/tests/test_array.py
@@ -14,7 +14,7 @@ from __future__ import absolute_import
 import unittest
 
 from traits.api import Array, Bool, HasTraits
-from traits.testing.import_handler import numpy, requires_numpy
+from traits.testing.optional_dependencies import numpy, requires_numpy
 
 
 if numpy is not None:
