+++ traits/tests/test_bool.py
@@ -19,7 +19,7 @@ import unittest
 import six
 
 from traits.api import Bool, Dict, HasTraits, Int, TraitError
-from traits.testing.import_handler import numpy, requires_numpy
+from traits.testing.optional_dependencies import numpy, requires_numpy
 
 if six.PY2:
     LONG_TYPE = long
