+++ traits/tests/test_cythonized_traits.py
@@ -13,7 +13,7 @@ The tests need a Cython version > 0.19 and a compiler.
 import unittest
 
 from traits.testing.unittest_tools import UnittestTools
-from traits.testing.import_handler import cython
+from traits.testing.optional_dependencies import cython
 
 
 def has_no_compiler():
