+++ traits/tests/test_integer.py
@@ -25,7 +25,7 @@ import unittest
 import six
 
 from traits.api import HasTraits, Int, TraitError
-from traits.testing.import_handler import numpy, requires_numpy
+from traits.testing.optional_dependencies import numpy, requires_numpy
 
 if six.PY2:
     LONG_TYPE = long
